exp4/requirements.txt
numpy
matplotlib
tqdm
scipy
torchmetrics
pytorch-fid

